feat: abstraction over inference provider shared between podman and kubernetes
