+++ packages/backend/src/workers/provider/InferenceProvider.spec.ts
@@ -18,7 +18,7 @@
 
 import { beforeEach, describe, expect, test, vi } from 'vitest';
 import type { TaskRegistry } from '../../registries/TaskRegistry';
-import { type BetterContainerCreateResult, InferenceProvider } from './InferenceProvider';
+import { type BetterContainerCreateResult, PodmanInferenceProvider } from './PodmanInferenceProvider';
 import type { InferenceServerConfig } from '@shared/src/models/InferenceServerConfig';
 import type {
   ContainerCreateOptions,
@@ -61,7 +61,7 @@ const taskRegistry: TaskRegistry = {
   updateTask: vi.fn(),
 } as unknown as TaskRegistry;
 
-class TestInferenceProvider extends InferenceProvider {
+class TestInferenceProvider extends PodmanInferenceProvider {
   constructor() {
     super(taskRegistry, InferenceType.NONE, 'test-inference-provider');
   }
